Frequency Band of all Channels
Based on the earlier research, I set the EEG frequency band to the variable "eeg_bands" to determine the appropriate frequencies, the FFT is computed on the EEG data. Finally, the EEG powers for each channel are presented as a "Bar chart" after the band power is determined by adding the squared FFT values within the chosen frequency range and normalizing by the length of FFT values.